+++ app/perf/rfc4175_422be10_to_y210.c
@@ -6,8 +6,8 @@
 
 static int perf_cvt_422_10_pg2_be_to_y210(mtl_handle st, int w, int h, int frames,
                                           int fb_cnt) {
-  size_t fb_pg2_size = w * h * 5 / 2;
-  size_t fb_pg2_size_y210 = w * h * 4;
+  size_t fb_pg2_size = (size_t)w * h * 5 / 2;
+  size_t fb_pg2_size_y210 = (size_t)w * h * 4;
   mtl_udma_handle dma = mtl_udma_create(st, 128, MTL_PORT_P);
   struct st20_rfc4175_422_10_pg2_be* pg_be =
       (struct st20_rfc4175_422_10_pg2_be*)mtl_hp_malloc(st, fb_pg2_size * fb_cnt,
