+++ app/perf/rfc4175_422be12_to_p12le.c
@@ -6,14 +6,14 @@
 
 static int perf_cvt_422_12_pg2_to_planar_le(mtl_handle st, int w, int h, int frames,
                                             int fb_cnt) {
-  size_t fb_pg2_size = w * h * 6 / 2;
+  size_t fb_pg2_size = (size_t)w * h * 6 / 2;
   mtl_udma_handle dma = mtl_udma_create(st, 128, MTL_PORT_P);
   struct st20_rfc4175_422_12_pg2_be* pg_be =
       (struct st20_rfc4175_422_12_pg2_be*)mtl_hp_malloc(st, fb_pg2_size * fb_cnt,
                                                         MTL_PORT_P);
   mtl_iova_t pg_be_iova = mtl_hp_virt2iova(st, pg_be);
   mtl_iova_t pg_be_in_iova;
-  size_t planar_size = w * h * 2 * sizeof(uint16_t);
+  size_t planar_size = (size_t)w * h * 2 * sizeof(uint16_t);
   float planar_size_m = (float)planar_size / 1024 / 1024;
   uint16_t* p12_u16 = (uint16_t*)malloc(planar_size * fb_cnt);
   uint16_t* p12_u16_b = p12_u16 + w * h;
